fix: Update rand::random usage for rand 0.9 compatibility
- Change rand::random::<usize>() to rand::random::<u32>() as usize
- rand 0.9 removed support for usize in StandardUniform distribution
- All tests passing

+++ examples/api_polling_scraping.rs
@@ -100,7 +100,7 @@ impl DataStore {
         let data = WeatherData {
             city: city.to_string(),
             temperature: 15.0 + (rand::random::<f64>() * 20.0),
-            conditions: vec!["Sunny", "Cloudy", "Rainy", "Partly Cloudy"][rand::random::<usize>() % 4].to_string(),
+            conditions: vec!["Sunny", "Cloudy", "Rainy", "Partly Cloudy"][(rand::random::<u32>() % 4) as usize].to_string(),
             timestamp: Utc::now(),
         };
